+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-**4.1.1 - 04/17/26**
+**4.1.1 - 04/21/26**
 
 - Raise error if registering duplicate Pipelines